+++ src/main/java/dev/doctor4t/ratatouille/client/model/armor/CustomArmorModelDefinition.java
@@ -1,0 +1,10 @@
+package dev.doctor4t.ratatouille.client.model.armor;
+
+import net.minecraft.client.model.ModelData;
+import net.minecraft.util.Identifier;
+
+public abstract class CustomArmorModelDefinition {
+    public abstract void addModelParts(ModelData modelData);
+
+    public abstract Identifier getTexture();
+}